Remove no longer needed json headers. Refactor json_utils
