+++ minomaly/CMakeLists.txt
@@ -81,7 +81,6 @@ add_library(${_MINO_ENGINE} ${PROJECT_SOURCE_DIR}/engine/window/window.cpp
 add_library(${_MINO_COMMON} ${PROJECT_SOURCE_DIR}/common/surface/surface.cpp
                             ${PROJECT_SOURCE_DIR}/common/texture/texture.cpp
                             ${PROJECT_SOURCE_DIR}/common/bounding_box/bounding_box.cpp
-                            ${PROJECT_SOURCE_DIR}/common/json/json_serializer.cpp
                             ${PROJECT_SOURCE_DIR}/common/font/font.cpp
                             ${PROJECT_SOURCE_DIR}/common/audio/audio.cpp)
