+++ minomaly/common/mino_common.h
@@ -2,7 +2,7 @@
 #include "audio.h"
 #include "color.h"
 #include "font.h"
-#include "json_serializer.h"
+#include "json.h"
 #include "matrix.h"
 #include "observer.h"
 #include "quadtree.h"
